requirements.txt
boto3==1.1.4
botocore==1.2.5
pandas==0.17.1
psycopg2==2.6.1
python-dotenv==0.1.5
chardet==2.3.0
